Add migration code for existing ClusterDeployments
Since this patch set changes the way this operator works to reconcile
a new custom resource, we need to be able to get existing
ClusterDeployments with their PagerDuty services into this new
format.

Rather than adding some sort of additional field to the spec that
would only ever be useful once for one CR, this looks for the
following annotation on CRDs: "pd.openshift.io/legacy" with a
non-empty value. This allows us to set this annotation on the OSD
PagerDutyIntegration CR, so that the operator knows that this is the
one to use for migration.

Once it has migrated all of the resources, it will delete the
annotation from the PagerDutyIntegration CR, so that it won't attempt
to migrate on every reconciliation.